refactor and add concepts-results
